refactor: use toTile instead of geoCoordsToTile
