+++ src/tiles/tiles.ts
@@ -276,46 +276,3 @@ export function tileEffectiveWidth(tileMatrix: TileMatrix): number {
   const { cellSize, tileWidth } = tileMatrix;
   return cellSize * tileWidth;
 }
-
-/**
- * Calculates a tile for a longitude, latitude and tile matrix
- * @param geoPoint point with longitude and latitude
- * @param tileMatrix tile matrix which the calculated tile belongs to
- * @param reverseIntersectionPolicy boolean value whether to reverse the intersection policy (in cases that the location is on the edge of the tile)
- * @param metatile size of a metatile
- */
-export function geoCoordsToTile<T extends TileMatrixSet>(
-  geoPoint: GeoPoint,
-  tileMatrix: ArrayElement<T['tileMatrices']>,
-  reverseIntersectionPolicy: boolean,
-  metatile = 1
-): Tile<T> {
-  const width = tileEffectiveWidth(tileMatrix) * metatile;
-  const height = tileEffectiveHeight(tileMatrix) * metatile;
-
-  const { min: tileMatrixBoundingBoxMin, max: tileMatrixBoundingBoxMax } = tileMatrixToBoundingBox(tileMatrix);
-
-  const {
-    identifier: { code: tileMatrixId },
-    cornerOfOrigin = 'topLeft',
-  } = tileMatrix;
-
-  const x = (geoPoint.lon - tileMatrixBoundingBoxMin.lon) / width;
-  const y = (cornerOfOrigin === 'topLeft' ? tileMatrixBoundingBoxMax.lat - geoPoint.lat : geoPoint.lat - tileMatrixBoundingBoxMin.lat) / height;
-
-  // When explicitly asked to reverse the intersection policy (location on the edge of the tile)
-  if (reverseIntersectionPolicy) {
-    const tileCol = Math.ceil(x) - 1;
-    const tileRow = Math.ceil(y) - 1;
-    return new Tile(tileCol, tileRow, tileMatrixId, metatile);
-  }
-
-  // When longitude/latitude is on the maximum edge of the tile matrix (e.g. lon = 180 lat = 90)
-  const onEdgeXTranslation = geoPoint.lon === tileMatrixBoundingBoxMax.lon ? 1 : 0;
-  const onEdgeYTranslation = geoPoint.lat === (cornerOfOrigin === 'topLeft' ? tileMatrixBoundingBoxMin.lat : tileMatrixBoundingBoxMax.lat) ? 1 : 0;
-
-  const tileCol = Math.floor(x) - onEdgeXTranslation;
-  const tileRow = Math.floor(y) - onEdgeYTranslation;
-
-  return new Tile(tileCol, tileRow, tileMatrixId, metatile);
-}
